requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
openai==1.35.13
pymongo==4.8.0